refactor: remove githubOrganization and use githubIntegrationService
